Do not let muxing fail silently
If the muxing process exits with non-zero status code, it would be useful to deal with that error. As of now it fails but the Python interpreter doesn't know it.
To give an example, tried muxing a file with a non-existing, non-user-writable output path and the process failed but the interpreter didn't catch any error, so it seemed like it worked when in fact it didn't.