Add support to Path objects
MKVTrack and MKVFile().mux() already support the file_path argument as a Path object, it would be great to add this support everywhere, instead of converting it every time with str(). For instance, it would be even better to add support for general os.path-like objects, not only pathlib's Path, what do you think?