Release/1.0.8 (#63)
* Do not let muxing fail silently

If the muxing process exits with non-zero status code, it would be useful to deal with that error. As of now it fails but the Python interpreter doesn't know it.
To give an example, tried muxing a file with a non-existing, non-user-writable output path and the process failed but the interpreter didn't catch any error, so it seemed like it worked when in fact it didn't.

* Add support for os.PathLike objects (#62)

* Add support for Path objects

MKVTrack and MKVFile's mux() already support passing a Path object for the file_path argument. Would be even better to add support for general os.path-like object, not only strings.

* Update Verifications.py

Added support for all Path-like objects, instead of only pathlib's Path.

Co-authored-by: Marco Rubin <20150305+Rubo3@users.noreply.github.com>